+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 * Assign Suites button still visible when no possible suites to assign ([#976](https://gitlab.com/yale-sdmp/vesta/issues/976)).
 * [Admin Dashboard] Add spec for draws create ([#1006](https://gitlab.com/yale-sdmp/vesta/issues/1006)).
 * [Admin Dashboard] Add spec for draws index ([#1003](https://gitlab.com/yale-sdmp/vesta/issues/1003)).
+* [Admin Dashboard] Add spec for colleges index, show, create, update, and destroy ([#1010](https://gitlab.com/yale-sdmp/vesta/issues/1010)).
 
 ### Changed
 * Change form cancel operation to standardize redirects to page referrer ([#979](https://gitlab.com/yale-sdmp/vesta/issues/979)).